Optimize array copying in loadProgram calls
Instead of copying to a new slice every time, keep the same backing
arrays when possible and change sizes as needed. This drasically lowers
the amount of memory being allocated and deallocating for over 100x
speed increses on this opcode.

This opcode turned out to be used more than expected, so more than a
trivial solution was required. Now can run sandmark in less than 4.5
minutes instead of hours.